// src/store/score/types.ts
export enum ScoreActionTypes {
  FETCH_SCORE_REQUEST = "@@SCORE/FETCH/SCORE_REQUEST",
  FETCH_SCORE_ERROR = "@@SCORE/FETCH/SCORE_ERROR",
  FETCH_SCORE_SUCCESS = "@@SCORE/FETCH/SCORE_SUCCESS",

  FETCH_BUYITEM_REQUEST = "@@SCORE/FETCH/BUYITEM_REQUEST",
  FETCH_BUYITEM_SUCCESS = "@@SCORE/FETCH/BUYITEM_SUCCESS",

  FETCH_BUYLEVELMED_REQUEST = "@@SCORE/FETCH/BUYLEVELMED_REQUEST",
  FETCH_BUYLEVELMED_SUCCESS = "@@SCORE/FETCH/BUYLEVELMED_SUCCESS",

  FETCH_BUYLEVELHARD_REQUEST = "@@SCORE/FETCH/BUYLEVELHARD_REQUEST",
  FETCH_BUYLEVELHARD_SUCCESS = "@@SCORE/FETCH/BUYLEVELHARD_SUCCESS",
}